requirements.txt
yfinance
nsetools
pandas
sqlalchemy
twilio
python-dotenv
apscheduler
pytz
